Latest build

+++ src/functions/public/Get-WellKnownSid.ps1
@@ -21,24 +21,55 @@ function Get-WellKnownSid {
     [OutputType([Microsoft.Management.Infrastructure.CimInstance])]
     param (
         [Parameter(ValueFromPipeline)]
-        [string[]]$CimServerName
+        [string[]]$CimServerName,
+
+        # Cache of known Win32_Account instances keyed by domain and SID
+        [hashtable]$Win32AccountsBySID = ([hashtable]::Synchronized(@{})),
+
+        # Cache of known Win32_Account instances keyed by domain (e.g. CONTOSO) and Caption (NTAccount name e.g. CONTOSO\User1)
+        [hashtable]$Win32AccountsByCaption = ([hashtable]::Synchronized(@{})),
+
+        [hashtable]$DomainsByFqdn = ([hashtable]::Synchronized(@{})),
+
+        # Cache of known directory servers to reduce duplicate queries
+        [hashtable]$AdsiServersByDns = [hashtable]::Synchronized(@{}),
+
+        [hashtable]$DomainsByNetbios = ([hashtable]::Synchronized(@{}))
     )
+    begin {
+        $AdsiServersWhoseWin32AccountsExistInCache = $Win32AccountsBySID.Keys |
+        ForEach-Object { ($_ -split '\\')[0] } |
+        Sort-Object -Unique
+    }
     process {
         ForEach ($ThisServer in $CimServerName) {
             if ($ThisServer -eq (hostname) -or $ThisServer -eq 'localhost' -or $ThisServer -eq '127.0.0.1' -or [string]::IsNullOrEmpty($ThisServer)) {
-                Write-Debug -Message "  $(Get-Date -Format s)`t$(hostname)`tGet-WellKnownSid`t`$CimSession = New-CimSession"
-                Write-Debug -Message "  $(Get-Date -Format s)`t$(hostname)`tGet-WellKnownSid`tGet-CimInstance -ClassName Win32_Account -CimSession `$CimSession"
-                $CimSession = New-CimSession
                 $ThisServer = hostname
-            } else {
-                Write-Debug -Message "  $(Get-Date -Format s)`t$(hostname)`tGet-WellKnownSid`t`$CimSession = New-CimSession -ComputerName '$ThisServer'"
-                Write-Debug -Message "  $(Get-Date -Format s)`t$(hostname)`tGet-WellKnownSid`tGet-CimInstance -ClassName Win32_Account -CimSession `$CimSession"
-                $CimSession = New-CimSession -ComputerName $ThisServer
             }
+            # Return matching objects from the cache if possible rather than performing a CIM query
+            # The cache is based on the Caption of the Win32 accounts which conatins only NetBios names
+            $ThisServerNetbios = ConvertTo-LDAPDomainNetBIOS -DomainFQDN
+            if ($AdsiServersWhoseWin32AccountsExistInCache -contains $ThisServer) {
+                $Win32AccountsBySID.Keys | ForEach-Object {
+                    if ($_ -like "$ThisServer\*") {
+                        $Win32AccountsBySID[$_]
+                    }
+                }
+            } else {
+                if ($ThisServer -eq (hostname)) {
+                    Write-Debug -Message "  $(Get-Date -Format s)`t$(hostname)`tGet-WellKnownSid`t`$CimSession = New-CimSession # For '$ThisServer'"
+                    $CimSession = New-CimSession
+                    Write-Debug -Message "  $(Get-Date -Format s)`t$(hostname)`tGet-WellKnownSid`tGet-CimInstance -ClassName Win32_Account -CimSession `$CimSession # For '$ThisServer'"
+                } else {
+                    Write-Debug -Message "  $(Get-Date -Format s)`t$(hostname)`tGet-WellKnownSid`t`$CimSession = New-CimSession -ComputerName '$ThisServer' # For '$ThisServer'"
+                    $CimSession = New-CimSession -ComputerName $ThisServer
+                    Write-Debug -Message "  $(Get-Date -Format s)`t$(hostname)`tGet-WellKnownSid`tGet-CimInstance -ClassName Win32_Account -CimSession `$CimSession # For '$ThisServer'"
+                }
 
-            Get-CimInstance -ClassName Win32_Account -CimSession $CimSession
+                Get-CimInstance -ClassName Win32_Account -CimSession $CimSession
 
-            Remove-CimSession -CimSession $CimSession
+                Remove-CimSession -CimSession $CimSession
+            }
         }
     }
 }
